Relocate media router package
