+++ website/index.html
@@ -97,7 +97,7 @@ assertThat(layout.getShowDividers()).isEqualTo(SHOW_DIVIDERS_MIDDLE);</pre>
           <li><p>Google Play Services</p><pre class="prettyprint">import static com.pkware.truth.android.playservices.Assertions.assertThat;</pre></li>
           <li><p>appcompat</p><pre class="prettyprint">import static com.pkware.truth.androidx.appcompat.Assertions.assertThat;</pre></li>
           <li><p>Material Components</p><pre class="prettyprint">import static com.pkware.truth.android.material.Assertions.assertThat;</pre></li>
-          <li><p>mediarouter</p><pre class="prettyprint">import static com.pkware.truth.android.mediarouter.v7.Assertions.assertThat;</pre></li>
+          <li><p>mediarouter</p><pre class="prettyprint">import static com.pkware.truth.androidx.mediarouter.Assertions.assertThat;</pre></li>
           <li><p>gridlayout</p><pre class="prettyprint">import static com.pkware.truth.androidx.gridlayout.Assertions.assertThat;</pre></li>
           <li><p>cardview</p><pre class="prettyprint">import static com.pkware.truth.androidx.cardview.Assertions.assertThat;</pre></li>
           <li><p>recyclerview</p><pre class="prettyprint">import static com.pkware.truth.android.recyclerview.v7.Assertions.assertThat;</pre></li>
